+++ apollo-server-app/src/controllers/TestController.js
@@ -1,9 +1,0 @@
-import resolvers from './../resolvers/TestResolver';
-import testType from '../typeDefs/TestSchema';
-import SrmAPI from '../models/ApiSrmModel';
-
-export {
-  SrmAPI,
-  resolvers as TestResolver,
-  testType as TestSchema,
-};