+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Mutation newEnterpriseProvider[(29)]
 - Add Mutation asssignRepresentativeToProviderEnterprise[(29)]
 - Add Mutation setRepresentativeFavoriteProvider[(29)]
+- Add query getPayerListFromProvider[(30)]
+- Add query getProviderListFromPayer[(30)]
 
 ### Changed
 - Change input of asssignRepresentativeToPayerEnterprise service [(29)]
@@ -97,3 +99,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [(25)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/25/
 [(26)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/26/
 [(29)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/29/
+[(30)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/30/